requirements.txt
# review-mind-rag 의존성
# RAG 기반 이커머스 리뷰 분석 시스템

# === LLM Framework ===
langchain==0.2.16
langchain-openai==0.1.25
langchain-chroma==0.1.4
langchain-community==0.2.17

# === Vector Database ===
chromadb>=0.5.0

# === OpenAI ===
openai>=1.50.0

# === Web UI ===
streamlit>=1.40.0

# === Data Processing ===
pandas>=2.0.0
numpy>=1.24.0

# === Visualization ===
plotly>=5.18.0
matplotlib>=3.8.0
wordcloud>=1.9.0

# === Utilities ===
python-dotenv>=1.0.0
tqdm>=4.66.0
requests>=2.31.0

# === Data Loading ===
datasets>=2.14.0  # Hugging Face datasets for Amazon Reviews

# === Reranking ===
sentence-transformers>=2.2.0  # Cross-encoder for reranking

# === Testing ===
pytest>=8.0.0
pytest-cov>=4.1.0
pytest-asyncio>=0.23.0
pytest-mock>=3.12.0

# === E2E Testing ===
playwright>=1.40.0
pytest-playwright>=0.4.0
